chatbox at least exists now, but just as a menu item and is not doing anything when clicked
